scripts/requirements.txt
opencv-python
numpy
requests
paho-mqtt
